+++ OrchestratorV2/interfaces/helpers/buildhelper.ts
@@ -2,6 +2,6 @@ import { Build } from "azure-devops-node-api/interfaces/BuildInterfaces";
 
 export interface IBuildHelper {
 
-    findBuild(projectName: string, definitionName: string, definitionId: number, top: number, tags: string[]): Promise<Build>;
+    findBuild(projectName: string, definitionName: string, definitionId: number, buildNumber: string, tags: string[], top: number): Promise<Build>;
 
 }