+++ OrchestratorV2/interfaces/task/filters.ts
@@ -2,6 +2,7 @@ export interface IFilters {
 
     releaseTags: string[];
     artifactTags: string[];
+    artifactVersion: string;
     artifactBranch: string;
     stageStatuses: string[];