+++ OrchestratorV2/orchestrator/filtrator.ts
@@ -44,14 +44,24 @@ export class Filtrator implements IFiltrator {
             let artifactBranch: string | undefined;
 
             // Get build artifact version
-            if (filters.artifactTags.length) {
+            if (filters.artifactVersion || filters.artifactTags.length) {
 
-                debug(`Using <${String.Join("|", filters.artifactTags)}> artifact tag filter`);
+                if (filters.artifactVersion) {
+
+                    debug(`Using <${filters.artifactVersion}> artifact version filter`);
+
+                }
+
+                if (filters.artifactTags.length) {
+
+                    debug(`Using <${String.Join("|", filters.artifactTags)}> artifact tag filter`);
+
+                }
 
                 const buildDefinitionName: string = primaryBuildArtifact.definitionReference!.definition.name!;
                 const buildDefinitionId: number = Number(primaryBuildArtifact.definitionReference!.definition.id!);
 
-                const buildArtifact: Build = await this.buildHelper.findBuild(project.name!, buildDefinitionName, buildDefinitionId, 100, filters.artifactTags);
+                const buildArtifact: Build = await this.buildHelper.findBuild(project.name!, buildDefinitionName, buildDefinitionId, filters.artifactVersion, filters.artifactTags, 100);
 
                 artifactVersion = buildArtifact.id!.toString();
 
@@ -97,14 +107,24 @@ export class Filtrator implements IFiltrator {
         if (primaryBuildArtifact) {
 
             // Add build artifact version
-            if (filters.artifactTags.length) {
+            if (filters.artifactVersion || filters.artifactTags.length) {
 
-                debug(`Using <${String.Join("|", filters.artifactTags)}> artifact tag filter`);
+                if (filters.artifactVersion) {
+
+                    debug(`Using <${filters.artifactVersion}> artifact version filter`);
+
+                }
+
+                if (filters.artifactTags.length) {
+
+                    debug(`Using <${String.Join("|", filters.artifactTags)}> artifact tag filter`);
+
+                }
 
                 const buildDefinitionName: string = primaryBuildArtifact.definitionReference!.definition.name!;
                 const buildDefinitionId: number = Number(primaryBuildArtifact.definitionReference!.definition.id!);
 
-                const buildArtifact: Build = await this.buildHelper.findBuild(project.name!, buildDefinitionName, buildDefinitionId, 100, filters.artifactTags);
+                const buildArtifact: Build = await this.buildHelper.findBuild(project.name!, buildDefinitionName, buildDefinitionId, filters.artifactVersion, filters.artifactTags, 100);
 
                 releaseFilter.artifactVersionId = buildArtifact.id!;