+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Major code refactoring & improvements
 - Update documentation & add privacy policy
 - Update task to use name parameters ([4](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/4), [43](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/43))
+- Add artifact version filter support ([47](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/47))
 - Add latest release stage status filter support ([26](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/26))
 - Add release variables support ([24](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/24))
 - Make approval retries configurable ([23](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/23))